==> Pacbio_R/7_IsoSeq_transcript_TPM.Rmd <==
---
title: "differential transcript usage"
author: "Kelly Ren"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  bookdown::html_document2: default
bibliography: FinalthesisRef.bib
link-citations: yes
fig_caption: yes
toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

```{r}
library(magrittr)
library(dplyr)
library(stringr)
library(readr)
library(tibble)
library(DRIMSeq)
library(tidyr)
library(ggplot2)
'%!in%' <- function(x,y){!('%in%'(x,y))}
```

# read the count data
```{r}
associated_isoform_counts_with_novel_prefilter <- readRDS("associated_isoform_counts_with_novel_prefilter.rds")
```

```{r}
# associated_isoform_TPM_with_novel_prefilter <- associated_isoform_counts_with_novel_prefilter %>%
#   mutate(FL22_TPM = (FL22/(sum(FL22)+sum(FL99)+sum(FL65)+sum(FL53)+sum(FL60)+sum(FL60)))*1e6,
#          FL99_TPM = (FL99/(sum(FL22)+sum(FL99)+sum(FL65)+sum(FL53)+sum(FL60)+sum(FL60)))*1e6,
#          FL65_TPM = (FL65/(sum(FL22)+sum(FL99)+sum(FL65)+sum(FL53)+sum(FL60)+sum(FL60)))*1e6,
#          FL53_TPM = (FL53/(sum(FL22)+sum(FL99)+sum(FL65)+sum(FL53)+sum(FL60)+sum(FL60)))*1e6,
#          FL60_TPM = (FL60/(sum(FL22)+sum(FL99)+sum(FL65)+sum(FL53)+sum(FL60)+sum(FL60)))*1e6,
#          FL7_TPM = (FL7/(sum(FL22)+sum(FL99)+sum(FL65)+sum(FL53)+sum(FL60)+sum(FL60)))*1e6)


associated_isoform_TPM_with_novel_prefilter <- associated_isoform_counts_with_novel_prefilter %>%
  mutate(FL22_TPM = (FL22/(sum(FL22)))*1e6,
         FL99_TPM = (FL99/(sum(FL99)))*1e6,
         FL65_TPM = (FL65/(sum(FL65)))*1e6,
         FL53_TPM = (FL53/(sum(FL53)))*1e6,
         FL60_TPM = (FL60/(sum(FL60)))*1e6,
         FL7_TPM = (FL7/(sum(FL7)))*1e6)

associated_isoform_TPM_with_novel_prefilter <- associated_isoform_TPM_with_novel_prefilter[,c("FL22_TPM","FL99_TPM","FL65_TPM","FL53_TPM","FL60_TPM","FL7_TPM")]

saveRDS(associated_isoform_TPM_with_novel_prefilter,"associated_isoform_TPM_with_novel_prefilter.rds")
```

```{r}
associated_isoform_counts_with_novel_afterfilter <- readRDS("associated_isoform_counts_with_novel_afterfilter.rds")

associated_isoform_TPM_with_novel_prefilter[rownames(associated_isoform_counts_with_novel_afterfilter),] %>%
  saveRDS("associated_isoform_TPM_with_novel_afterfilter.rds")
```

# sessionInfo
```{r}
sessionInfo()
```
